[Commit 12] Correcting Blanck searched word issue; creating extensions group
